Move test to correct file
